[Rename] 컴포넌트 및 파일명 오타 수정 - finsh -> finish
